split quota manager in two compilation units
