+++ cvmfs/talk.cc
@@ -44,7 +44,7 @@
 #include "nfs_maps.h"
 #include "options.h"
 #include "platform.h"
-#include "quota.h"
+#include "quota_posix.h"
 #include "shortstring.h"
 #include "statistics.h"
 #include "tracer.h"
